refactor: extract cardFront to submodule #37
